rebase and initial clean up
