+++ app/services/question_service.js
@@ -9,6 +9,17 @@
     }
 */
 
+const { extractValidFields, validateAgainstSchema } = require('../../lib/validator')
+
+const questionInformationSchema = {
+    id: {required: true},
+    courseId: {required: true},
+    type: {required: true},
+    stem: {required: true},
+    content: {required: false},
+    answers: {required: false}
+}
+
 const getQuestionScore = function (question, submission) {
     let grade = 0
     switch(question.type) {
